requirements.txt
# Core dependencies
streamlit>=1.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
mysql-connector-python>=8.0.0

# Development dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.0.0
mypy>=1.5.0
